FIX: Mutate_AddConnection() for feedforward networks did not evenly distribute the selection of random source and target neurons. This could result in some neurons having zero probability.
FIX: Mutate_ConnectionWeights() did not evenly distribute the selection of connections to mutate.

FIX/ENHANCEMENT: Genome sorting now uses an unstable sort algorithm that randomly distributes the position of genomes of equal fitness. This removes the previous arbitrary selection of the top N genomes for elitism and selection. This might improve fostering and maintanining of genetic diversity.